Add CPE security tag for zlib and mark a vuln as known patched so it doesn't show up in scan results.
An osv-scanner scan of zlib:1.3.1 shows no open vulns as of now. While
this package is the chromium fork of zlib, The CPE tag (and license)
needs to reflect the original source for legal compliance reasons.

Adding CPE security tags to a METADATA file is necessary to enable
automated vulnerability monitoring.

Bug: 432520108
Change-Id: Icdcb06cc434c999d93bb6295d7fceb134b793c23